cleanup boost

+++ lib/gears/boost.rb
@@ -16,9 +16,13 @@ module Gears
 
     def obtain()
       # github_obtain('boostorg', 'boost')
+      Dir.chdir(_root_dir + '/build')
+      return true if Dir.exist? 'Boost'
+
       `wget http://downloads.sourceforge.net/project/boost/boost/1.59.0/boost_1_59_0.zip`
       `unzip boost_1_59_0.zip`
       `mv boost_1_59_0 Boost`
+      FileUtils.rm('boost_1_59_0.zip')
       true
     end
